scheduledjobs: move MaybeRewriteCronExpr into package
This was moved from the schematelemetrycontroller package.

There are no code changes in this commit.

Release note: None
