+++ pkg/BUILD.bazel
@@ -279,6 +279,7 @@ ALL_TESTS = [
     "//pkg/rpc/nodedialer:nodedialer_test",
     "//pkg/rpc:rpc_test",
     "//pkg/scheduledjobs/schedulebase:schedulebase_test",
+    "//pkg/scheduledjobs:scheduledjobs_test",
     "//pkg/security/certmgr:certmgr_test",
     "//pkg/security/password:password_test",
     "//pkg/security/sessionrevival:sessionrevival_test",
@@ -352,7 +353,6 @@ ALL_TESTS = [
     "//pkg/sql/catalog/resolver:resolver_test",
     "//pkg/sql/catalog/schemadesc:schemadesc_test",
     "//pkg/sql/catalog/schemaexpr:schemaexpr_test",
-    "//pkg/sql/catalog/schematelemetry/schematelemetrycontroller:schematelemetrycontroller_test",
     "//pkg/sql/catalog/schematelemetry:schematelemetry_test",
     "//pkg/sql/catalog/seqexpr:seqexpr_disallowed_imports_test",
     "//pkg/sql/catalog/seqexpr:seqexpr_test",
@@ -1495,6 +1495,7 @@ GO_TARGETS = [
     "//pkg/scheduledjobs/schedulebase:schedulebase",
     "//pkg/scheduledjobs/schedulebase:schedulebase_test",
     "//pkg/scheduledjobs:scheduledjobs",
+    "//pkg/scheduledjobs:scheduledjobs_test",
     "//pkg/security/certmgr:certmgr",
     "//pkg/security/certmgr:certmgr_test",
     "//pkg/security/certnames:certnames",
@@ -1664,7 +1665,6 @@ GO_TARGETS = [
     "//pkg/sql/catalog/schemaexpr:schemaexpr",
     "//pkg/sql/catalog/schemaexpr:schemaexpr_test",
     "//pkg/sql/catalog/schematelemetry/schematelemetrycontroller:schematelemetrycontroller",
-    "//pkg/sql/catalog/schematelemetry/schematelemetrycontroller:schematelemetrycontroller_test",
     "//pkg/sql/catalog/schematelemetry:schematelemetry",
     "//pkg/sql/catalog/schematelemetry:schematelemetry_test",
     "//pkg/sql/catalog/seqexpr:seqexpr",
