+++ pkg/scheduledjobs/rewrite_cron_expr_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package schematelemetrycontroller
+package scheduledjobs
 
 import (
 	"bufio"
