+++ pkg/scheduledjobs/testdata/cron_rewrites
