+++ pkg/sql/catalog/schematelemetry/schema_telemetry_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobstest"
+	"github.com/cockroachdb/cockroach/pkg/scheduledjobs"
 	"github.com/cockroachdb/cockroach/pkg/sql"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schematelemetry/schematelemetrycontroller"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/builtins/builtinconstants"
@@ -82,7 +83,7 @@ func TestSchemaTelemetrySchedule(t *testing.T) {
 
 	clusterID := s.ExecutorConfig().(sql.ExecutorConfig).NodeInfo.
 		LogicalClusterID()
-	exp := schematelemetrycontroller.MaybeRewriteCronExpr(clusterID, "@weekly")
+	exp := scheduledjobs.MaybeRewriteCronExpr(clusterID, "@weekly")
 	tdb.CheckQueryResultsRetry(t, qExists, [][]string{{exp, "1"}})
 	tdb.ExecSucceedsSoon(t, qSet)
 	tdb.CheckQueryResultsRetry(t, qExists, [][]string{{"* * * * *", "1"}})
