+++ pkg/upgrade/upgrades/BUILD.bazel
@@ -148,6 +148,7 @@ go_test(
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/roachpb",
+        "//pkg/scheduledjobs",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/security/username",
